Fix scrollmagic whitespace bug
There was a bug where when going directly from some projects to others
(like from Trunks to Fisqual) there would be a large amount of
whitespace below the glams and before the actual process starts.

I think this was happening because the Duration of the individual
headings in the tracker was calculated based on a document.
getelementbyid call, which may be null if the page hadn't fully
loaded yet. I added in an else... return 0 and it seems to work

¯\_(ツ)_/¯